% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_manifest.R
\name{format_manifest}
\alias{format_manifest}
\title{Format manifest files}
\usage{
format_manifest(
  path,
  output_path = "./data/",
  exclude_cell_files,
  internal = FALSE
)
}
\arguments{
\item{path}{path where cell values are stored}

\item{output_path}{path where the manifest should be stored}

\item{exclude_cell_files}{data frame of subject files of data to ignore
(used if uploads fail to generate a new manifest)}

\item{internal}{process internally only}
}
\description{
Files are used by the panopts CLI
inteface to batch upload large amounts of files
}
