% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subject_file.R
\name{read_subject_file}
\alias{read_subject_file}
\title{Read subject set file}
\usage{
read_subject_file(
  path = "~/Desktop/jungle-weather-subjects.csv",
  subject_set = 81979
)
}
\arguments{
\item{path}{path of a subject set file}

\item{subject_set}{subject set to use}
}
\description{
Used in screening the manifests
for already uploaded samples (if the
upload routine fails badly).
}
