% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_corrections.R
\name{validate_corrections}
\alias{validate_corrections}
\title{Validate manifest corrections}
\usage{
validate_corrections(
  files = "data-raw/format_1/format_1_batch_1_corrections.csv",
  path = "/backup/cobecore/zooniverse/format_1_batch_1/"
)
}
\arguments{
\item{files}{CSV with files to check}

\item{path}{location of the image (cell) data}
}
\description{
visually check corrected / added files
to manifest statements
}
